Add initial Jacobian-derived acceleration function
Currently causes dynamics to diverge immediately.
Need to check the force values.
The previous linearised force function has been saved to aM_forces_linear:
note that it linearises the acceleration with respect to angles and velocities,
but the intensity integrals have not been properly linearised

Add num_processes argument to load_essential_data, to be passed to the new extract_opt